+++ src/Ext/BuildingType/Body.h
@@ -99,6 +99,11 @@ public:
 
 		ValueableVector<bool> HasPowerUpAnim;
 
+		Nullable<int> NewEvaVoice_Index;
+		Valueable<int> NewEvaVoice_Priority;
+		Valueable<bool> NewEvaVoice_RecheckOnDeath;
+		NullableIdx<VoxClass> NewEvaVoice_InitialMessage;
+
 		ExtData(BuildingTypeClass* OwnerObject) : Extension<BuildingTypeClass>(OwnerObject)
 			, PowersUp_Owner { AffectedHouse::Owner }
 			, PowersUp_Buildings {}
@@ -161,6 +166,10 @@ public:
 			, BuildingRepairedSound {}
 			, Refinery_UseNormalActiveAnim { false }
 			, HasPowerUpAnim {}
+			, NewEvaVoice_Index {}
+			, NewEvaVoice_Priority { 0 }
+			, NewEvaVoice_RecheckOnDeath { false }
+			, NewEvaVoice_InitialMessage { }
 		{ }
 
 		// Ares 0.A functions