+++ src/Ext/Rules/Body.cpp
@@ -311,6 +311,9 @@ void RulesExt::ExtData::LoadBeforeTypeData(RulesClass* pThis, CCINIClass* pINI)
 	this->DistributeTargetingFrame.Read(exINI, GameStrings::General, "DistributeTargetingFrame");
 	this->DistributeTargetingFrame_AIOnly.Read(exINI, GameStrings::General, "DistributeTargetingFrame.AIOnly");
 
+	// Reading Ares section [EVATypes] at evamd.ini
+	LoadEvaVoices();
+
 	this->InfantryAutoDeploy.Read(exINI, GameStrings::General, "InfantryAutoDeploy");
 
 	// Section AITargetTypes
@@ -575,6 +578,7 @@ void RulesExt::ExtData::Serialize(T& Stm)
 		.Process(this->TintColorBerserk)
 		.Process(this->AttackMove_IgnoreWeaponCheck)
 		.Process(this->AttackMove_StopWhenTargetAcquired)
+		.Process(this->EVAIndexList)
 		.Process(this->Parasite_GrappleAnim)
 		.Process(this->InfantryAutoDeploy)
 		;
@@ -617,6 +621,49 @@ void RulesExt::ExtData::ReplaceVoxelLightSources()
 
 	if (needCacheFlush)
 		Game::DestroyVoxelCaches();
+}
+
+void RulesExt::ExtData::LoadEvaVoices()
+{
+	CCFileClass pEvamdFile("evamd.ini");
+
+	if (pEvamdFile.Exists() && pEvamdFile.Open(FileAccessMode::Read))
+	{
+		CCINIClass iniEva;
+		iniEva.ReadCCFile(&pEvamdFile, true);
+		iniEva.CurrentSection = nullptr;
+		iniEva.CurrentSectionName = nullptr;
+		const auto pEvaSection = "EVATypes";
+
+		if (iniEva.GetSection(pEvaSection))
+		{
+			this->EVAIndexList.clear();
+
+			// Default EVA voices
+			this->EVAIndexList.emplace_back(GameStrings::Allied);
+			this->EVAIndexList.emplace_back(GameStrings::Russian);
+			this->EVAIndexList.emplace_back(GameStrings::Yuri);
+
+			// New EVA voices due to a new Ares section in evamd.ini
+			const auto count = (std::size_t)iniEva.GetKeyCount(pEvaSection);
+
+			for (std::size_t i = 0; i < count; i++)
+			{
+				const auto pEvaKey = iniEva.GetKeyName(pEvaSection, i);
+
+				if (iniEva.ReadString(pEvaSection, pEvaKey, "", Phobos::readBuffer) > 0)
+				{
+					std::string buffer = Phobos::readBuffer;
+					bool found = std::find(this->EVAIndexList.begin(), this->EVAIndexList.end(), buffer) != this->EVAIndexList.end();
+
+					if (!found)
+						this->EVAIndexList.emplace_back(buffer);
+				}
+			}
+		}
+	}
+
+	pEvamdFile.Close();
 }
 
 // =============================